Add support for remote yaml (.ibart.yaml) definitions
With this we introduce the new command: run_local_definitions to be used
in job-definition files. The command is a True/False command where True
means that in addition to the remote file '<remote>/.ibart.yaml' it will
also run the stock configuration at the server, i.e., the files that are
located in the 'jobdefs/' folder. A False means that it will only run
the remote file '<remote>/.ibart.yaml'.

If the one submitting a pull request is an owner of the project, then
then '.ibart.yml' will picked up from his branch. If he is not an owner,
then '.ibart.yml' will picked up from upstream projects main branch
(typically 'master'). Non-owners shall never be able to provide
'.ibart.yaml', since that would give them the ability to run arbitrary
commands on the server which can lead to security issue and lots of
damage on the server.

Fixes #8

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>